Move all the other helper classes to a separate file
